+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CentOS-6 6.9 x86_64, Apache 2.2, PHP 5.3, PHP memcached 1.0, PHP APC 3.1.
 - Removes scmi; it's maintained [upstream](https://github.com/jdeathe/centos-ssh/blob/centos-6/src/usr/sbin/scmi).
 - Fixes local port value in scmi install template.
 - Adds use of readonly variables for constants.
+- Adds support for event server MPM in bootstrap.
 
 ### 1.10.0 - 2017-07-13